Drop zip_code scope from routes for now
Gets build passing. Keeps zip_code in the listings table so we may come
back to this routing solution again at some point but will need to fix a
number of path generations in, eg, templates and specs.

It may also turn out not to be a primary/necessary component for all
listing routes.