move the listing form into vue
Ryan and I paired on this. I wanted to learn about Vue, and that was a success. We also talked about integrating Rails and Vue and our options.

This commit starts moving the form from ERB into Vue. The goal is to keep the functionality the same, which this doesn't quite do yet as the fields are being populated like they do on the Rails side. The next commits will get us there.

Ryan and I talked about whether it's worth spending the time reimplementing in Vue features that Rails already provides.

Upsides:

* Ryan knows Vue and I know Rails and we have time to pair on this, so know is a good time to establish a patterns that work to integrate the two that other people can then use as an example when working independently on other features.
* The pattern of putting this work in Vue so that people who are more comfortable with Vue than Rails have a nice path to start helping is something that worked on https://github.com/rubyforgood/voices-of-consent/, so this gives us options on how to break out next week's tasks

Downsides:

* We're taking time to reimplement things that already work

Other considerations:

* This gives us flexibility in dynamically reordering the form like we've discussed. That said, that's a feature we weren't planning to implement immediately, and could have postponed

Co-authored-by: Ryan Dowd <Ry-Dowd@users.noreply.github.com>