Allow `yarn test` to run alongside `bin/webpack-dev-server`
Currently, running tests will break the dev-server because the testing
build chain writes to the same `manifest.json` file used by the
dev-server, but with incompatible pack paths.

This configures the test env to not write the manifest to disk.

I'm not sure if there are other unintended side-effects to this change.
An alternative if this doesn't pan out is to instead change the
`publicPath` the manifest is written out to.