Create a default user
The `entrypoint.sh` script has been updated to create the default user
as defined in `db/seeds.rb`. This is idential to the result of executing
```bash
rake db:setup
rake db:seed
```

However, it will only happen if the database has not already been
initialize.

If the database has already been initialized (ie, if this is not the
first time this container has been launched), then the application will
attempt to migrate the database by calling `rake db:migrate`.

You can reset the container by running
```bash
docker-compose down --volumes --remove-orphans
```
which destroys all the containers and all the persistant state held in
the docker volumes attached to them. The database and the rest of the
application will be reinitialized from scratch the next time you run
```bash
docker-compose up --build
```

This commit also refactors the Dockerfile-webapp file into
webapp/Dockerfile, which will help IDEs and editors choose the right
syntax highlighting for that file.