Adds a README section detailing the use of docker
This is a pretty basic set of instructions for how docker can be used
with the currect version of the docker-compose setup. It should be
expanded at some point to describe how the various environment varibles
can be used to change and customize the behavior of the application, as
well as potentially giving more guidence to users who are not familiar
with docker and docker-compose.