Update the README with simpler setup instructions
When build the application in docker, it is nessesary to set a
SECRET_KEY_BASE environment variable so that rails is able to launch.
This can be generated using the `rake secret` command, and must be
placed into a `.env` file. Previously, the README assumed that the
user's host machine would have ruby installed - but in some cases, this
is not possible.

This allows the user to generate the .env file with a one line command.
It also uses the ruby install included in the `app` docker image, so
ruby does not need to be installed on the host system.