Introduce FormObjects to simplify our nested model forms
Currently, we're asking Interactors to do several things:
* Capture incoming params and apply them to nested objects
* Attempt to save nested objects, handling validation, transactions and
  error propogation
* If invalid, double for real objects during serialization

This is evidently too many responsibilities.

The intention here is to extract FormObjects whose sole responsibility
is to compose a (possibly nested) object graph and apply incoming
attributes. They are not responsible for validation or persistence.

Importantly, they _always_ return ActiveModel instances. This way, we
can use model instances for validation, persistence and serialization.

Note that in order to support partial updates, i've introduced the
concept of `given_inputs` on top of what ActiveInteraction provides.
This allows us to only apply given attributes on top of existing
objects, leaving other attributes as they are.