Add accepts_nested_attributes_for to help Submission creation
This helps AR propogate errors up from nested objects.

With this change, we end up using a hybrid of rails-way and custom
handling of nested objects.

In simpler cases, it may be simpler to just use AR:NestedAttributes.

In this particluar case, custom form objects were necessary because
we want to share some nested objects. Eg, Person and Listings should
point to the same Location while Listings and Submission point to
the same Person. This is logic that NestedAttributes would not handle.
OTOH, we still invoke it so that we don't have to hand-roll error
message propogation code.