Improve changelog format
Changes format to put issue, pr numbers _after_ the descriptive text.

While it was nice to have all the numbers aligned on the left, that
scheme breaks a bit when we want to list multiple issues/prs. Moreover,
i want to highlight _what_ the change was; the links are details that
viewers can dig into if they choose, not the primary subject.