Send category ids from ask/offer form instead of tag names
This sets us up to handle the situation where two subcategories might
have the same name, and allows us to ignore multiple nesting.

IMPORTANT: this introduces a bug: #576
We're intentionally accruing this debt because its not possible on
modern browsers to generate backend validation errors and fixing this
properly requires consolidating the `categories` and `tags` tables.

Co-authored-by: maebeale <maebeale@gmail.com>