WIP: Basic scaffold for CommunityResourceImporter
This creates a new CommunityResourceImporter class which can be used to import
CSV data and create CommunityResource objects and it's associated relationships.
At the moment this only creates the primary class and sets the name property.
There is still a lot more work to do, but what exists works and IT HAS TESTS!!!