Fix docker buildfile
Was resulting in an error when attempting to run the build:
  app_1    | /usr/local/bundle/gems/thor-1.0.1/lib/thor/parser/arguments.rb:133:in `parse_numeric': Expected numeric value for '--port'; got "${APP_PORT}" (Thor::MalformattedArgumentError)

There are two things that aren't quite set up right here:

The build arg was not in scope -- used args need to be _below_ the
`FROM` statement to be accessible.

CMD appears to not allow variable expansion, because it is processed at
runtime rather than build time.  Instead, we can export the `PORT`
envar, which will be picked up by `config/puma.rb` when starting rails.