Avoid duplicating records when importing community resources
When importing CommunityResource records, some rows in the dataset will have repeating
Organization or Community Resource information. When this happens we want to re-use the
existing objects in the database and update them accordingly. Specifically, instead of
duplicating a CommunityResource the importer will update it to include the new tag and
description.