Silence Ruby 2.7.0 deprecation warnings
This is intended temporarily silence deprecation warnings emitted from
third-party gems that have yet to become compatible with Ruby 2.7.0.
This is somewhat dangerous if left in long-term as warnings use of
soon-to-be-removed language features by the application itself will be
silenced.

This should be removed before the next time this code base upgrades
Ruby.

**NOTE**

This does NOT silence Rails deprecation warnings, these
should probably be addressed independently - a quick eyeball survey
shows that most of the Rails deprecation warnings are addressable in
the application code.

References
----------

- https://github.com/rubyforgood/mutual-aid/issues/718