Store gems on gem_cache volume
The docker-compose setup includes a persistent volume intended to
serve as a cache for bundler and rubygems. However, the volume mount
contains a small error which is causing gems to be installed on the
ephemeral filesyste - IE the gem cache is lost when a container
exits. This means that every request to start an app container needs
to re-install rubygems - and therefore re-run every step that comes
after in the Dockerfile. This leads to longer than necessary wait
times for the image to come up since the `bundle install` instruction
of the Docerfile will never produce a cache-able layer.

This fixes the path so that gems are stored on the persistent volume
and significantly decreases the amount of time one needs to wait for
app images and containers to become available.