Reduce the amount of work needed by docker build
In development, the app and webpacker services run in the same
context. Instead of building an image for each service, we now build one
image and share it - then let docker-compose figure out what to run,
which ports to expose, etc.

This uses Docker multi-stage builds to create the "base" image, then
the app and webpacker services both use the "base" image as a starting
point for their containers.

References
----------

- https://docs.docker.com/develop/develop-images/multistage-build/