Consolidate smtp config between dev and prod
All the relevant settings are configured via environment variables so we
can move towards a single config block in application.rb. This moves us
closer to 12factor, reducing the surface area for bugs and
misconfiguration.

Note the new `PORT` entry in `.env` which supports default_url_options
in mailers. This is pretty generic variable name which makes me a bit
nervous but was already being referenced in the app.