Improves db rake tasks for seeds
Uses db:reset which is faster to run, instead of running all migrations
to ensure that the schema is up to date. db:reset runs db:schema:load,
which ensures that everyone has the same db schema, so in case we need
to remove migrations, we don't rely on them for consistency.