Remove HtmlSanitizer.sanitize_for_vue
Found a bug with single-quotes in the custom regex we'd written and
while looking into this, realized we probably don't need a separate
sanitizer for vue. We were able to get the only current use of this
method (on the landing page) to work with sanitize_for_ralis along with
regular serialization (.to_json).

Co-authored-by: maebeale <maebeale@gmail.com>