Add Pagination for People
Closes #528

Using the [`pagy`](https://ddnexus.github.io/pagy/how-to#quick-start) gem,
add pagination to the `peoples_controller#index` route.

Two modules have been included:

  1. `Pagy::Backend` in `ApplicationController`
  2. `Pagy::FrontEnd` in `ApplicationHelper`

The navigation for `People` has been added to the `index` view and is
present only when there are enough `People` to paginate. `pagy` takes
care of all query parameters around pagination.

For testing, I introduced a new folder for controller testing and added
the `rails-controller-testing` gem to allow expectations via `assigns`.

I was not sure of any front-end specific testing for this feature but am
happy to create one if we need it.