Allow anonymous users to create comm logs
Integrating EmailNewSubmission into Asks|Offers Controllers broke those
forms for anonymous users. CommunicationLog.log_email allows a nil
`initiator` object but the DB does not allow null in the corresponding
`created_by_id` column.

It appears this column had a default id set to 1, but this would also
require persistence code to omit the `created_by` field if it isn't
provided; if a write is issued containing `created_by: nil`, rails will
pass that on to the DB which will reject it.

Moreover, i don't think it's wise to assume our 'system user' will always have
an id of 1. If we really want CommunicationLog entries to have a system
user as their default user id, we should explicitly set this in code.

Instead of implementing this additional logic, i've chosen here to drop
the default value and changed the column to be nullable. As such, `null`
can represent the same semantics of defaulting to the system user.