Drop shared pundit contexts
Inlining these snippets creates a smidge of duplication but i believe
this is a worthy tradeoff for simplicity, at least for now. In
particular, this eliminates the coupling amongst and between shared
contexts and specs via the shared `acting_user` let; such implicit
coupling can be surprising, hard to recognize, and cause hurdles
when refactoring.