Add can_admin? method to Context
While this duplicates a bit of logic in ApplicationPolicy, it didn't
feel quite right to have to do `policy(model).change?` in a
controller to determine where to redirect to. The `.change?` part of
that call seems an arbitrary choice. In fact, while it seems related to
permissions, in this case what i'm asking is not whether the user has
permissions to do something but whether they are acting as an admin.