Remove User#role
This presence of this method overrides the User's getter returning an instance of UserRole. However this doesn't work when trying to reference and display a role directly in the UI. e.g. #<UserRole:0x00007fa3f81dbd90> would be returned instead of the role name. Also, this role method is not being used anywhere in the codebase. If additional functionality is needed from UserRole in the future the class can be referenced directly