Use `string` filter instead of `record` for `location_type`
`active_interaction` doesn't coerce an empty string value to the default
value on `record` filters[1]. This causes the form to break on any form
where `location_type` is optional (eg CommunityResources#create).

We hadn't run into this on the Ask|Offer form because we only show the
location_type input once the user provides an address, and then rely on
the browser to enforce that it's populated.

[1] https://github.com/AaronLasseigne/active_interaction/issues/503