Style/StringLiterals: disable
Standard requires all string literals to use double-qoutes instead of
single-quotes.

This makes sense and is easily auto-corrected _but_ i'm not sure of
the value in enforcing this. The argument that using double-quotes saves
time (in trying to guess whether a string will need interpolation in the
future, or when that does change) is sound. However, any author who
ascribes to this is free to exclusively use double-quotes.

The downside of enforcing this is a significant disruption to git history
caused by this pervasive auto-correction. That in itself feels like
enough reason to just disable this rule.

https://docs.rubocop.org/rubocop/cops_style.html#stylestringliterals