Add to_boolean refinement
Uses an implementation that recognizes truthy strings from the set of
the following:
* Values recognized as true by YAML.load (https://yaml.org/type/bool.html).
* The opposites of values in ActiveRecord::Type::Boolean::FALSE_VALUES.
* Stripped, case-insensitive versions of both the above.

Importantly, there are functional differences from both YAML.load and
ActiveRecord. This implementation is an accept-list of truthy strings,
making all other strings false (except "" and nil, which return nil).

This is the opposite of AR casting, which is an accept-list of falsey
strings, and other strings are considered truthy. I feel it's usually more
important to strictly recognize truthy strings than falsey strings.  Eg,
params[:admin].to_boolean should only return true when it exactly matches
a known true value.

YAML.load has accept-lists for both truthy and falsey values, returning
a string for anything unrecognized, so this implementation differs in
defaulting unrecognized strings to false.

These differences are to be expected. The conversion from string to
boolean is very contextual and i suspect this is why there isn't a
standard `.to_b` implemented in ruby. I'm hoping the rules chosen here
are appropriate for our common use cases.