Fix bug in PersonPolicy that allows guest to access other People
`own_person` was implemented as `person.user_id == user&.id`. With a
guest user, where user is nil, this works but only if person has a user;
if it doesn't, then the two nils match and `own_person` returns true.

This effectively meant any guest user could access any Person record
that didn't have an associated user record. This is actually common
because listings can be created by guest, in which case we generate a
Person record w/o a user to associate it with.