Always apply HideRestrictedFieldsTransformer
Specifying a transform in a blueprint doesn't propogate down to its
views. This makes it onerous to do this explicitly, so switching to a
scheme where this transformer is applied to all blueprints by default.

The implementation then no-ops if a policy is not found or does not
impleement `restricted_attributes`.