Introduce BlueprintAuthorization module
`transform HideRestrictedFieldTransformer` at the top level of
a blueprint doesn't apply to any nested views. BlueprintAuthorization
overrides the `view` macro to explicitly apply the transformer.

With this change, the pattern for securing blueprints is to:
1. Implement a `restricted_attributes` method in the related policy.
2. `extend BlueprintAuthorization` in the blueprint.

Note that the previous commit went in a different direction but after
review and discussion, we decided on this approach.